LayoutTests/media/media-source/media-source-append-failed-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updatestart)
EVENT(error)
EVENT(updateend)
EVENT(sourceended)
END OF TEST
